+++ src/server/api/core/models.py
@@ -18,35 +18,64 @@ from langchain_community.embeddings.oci_generative_ai import OCIGenAIEmbeddings
 
 from giskard.llm.client.openai import OpenAIClient
 
-from server.utils.oci import init_genai_client
-from common.schema import ModelNameType, ModelTypeType, Model, ModelAccess, OracleCloudSettings
+from server.api.core import bootstrap, oci
+import common.schema as schema
 import common.logging_config as logging_config
 
-logger = logging_config.logging.getLogger("server.utils.models")
+logger = logging_config.logging.getLogger("api.core.models")
 
 
 #####################################################
 # Functions
 #####################################################
-async def apply_filter(
-    models_all: list[Model],
-    model_name: Optional[ModelNameType] = None,
-    model_type: Optional[ModelTypeType] = None,
-) -> list[Model]:
+def get_model(
+    model_name: Optional[schema.ModelNameType] = None,
+    model_type: Optional[schema.ModelTypeType] = None,
+) -> list[schema.Model]:
     """Used in direct call from list_models and agents.models"""
-    logger.debug("%i models are defined", len(models_all))
-    models_all = [
+    model_objects = bootstrap.MODEL_OBJECTS
+
+    logger.debug("%i models are defined", len(model_objects))
+
+    model_objects = [
         model
-        for model in models_all
+        for model in model_objects
         if (model_name is None or model.name == model_name) and (model_type is None or model.type == model_type)
     ]
-    logger.debug("%i models after filtering", len(models_all))
-    return models_all
+    logger.debug("%i models after filtering", len(model_objects))
+
+    if not model_objects:
+        raise ValueError(f"Model: {model_name} not found.")
+
+    return model_objects
 
 
-async def get_key_value(
-    model_objects: list[ModelAccess],
-    model_name: ModelNameType,
+def create_model(model: schema.Model) -> schema.Model:
+    """Create a new Model definition"""
+    model_objects = bootstrap.MODEL_OBJECTS
+    if any(d.name == model.name for d in model_objects):
+        raise ValueError(f"Model: {model.name} already exists.")
+
+    if not model.openai_compat:
+        openai_compat = next(
+            (model_config.openai_compat for model_config in model_objects if model_config.api == model.api),
+            False,
+        )
+        model.openai_compat = openai_compat
+    model_objects.append(model)
+
+    return get_model(model)
+
+
+def delete_model(name: schema.ModelNameType) -> None:
+    """Remove model from model objects"""
+    model_objects = bootstrap.MODEL_OBJECTS
+    bootstrap.MODEL_OBJECTS = [model for model in model_objects if model.name != name]
+
+
+def get_key_value(
+    model_objects: list[schema.ModelAccess],
+    model_name: schema.ModelNameType,
     model_key: str,
 ) -> str:
     """Return a models key value of its configuration"""
@@ -56,15 +85,15 @@ async def get_key_value(
     return None
 
 
-async def get_client(
-    model_objects: list[ModelAccess], model_config: dict, oci_config: OracleCloudSettings, giskard: bool = False
-) -> BaseChatModel:
+def get_client(model_config: dict, oci_config: schema.OracleCloudSettings, giskard: bool = False) -> BaseChatModel:
     """Retrieve model configuration"""
-    logger.debug("Model Config: %s; Giskard: %s", model_config, giskard)
+    logger.debug("schema.Model Config: %s; Giskard: %s", model_config, giskard)
+    model_objects = bootstrap.MODEL_OBJECTS
+
     model_name = model_config["model"]
-    model_api = await get_key_value(model_objects, model_name, "api")
-    model_api_key = await get_key_value(model_objects, model_name, "api_key")
-    model_url = await get_key_value(model_objects, model_name, "url")
+    model_api = get_key_value(model_objects, model_name, "api")
+    model_api_key = get_key_value(model_objects, model_name, "api_key")
+    model_url = get_key_value(model_objects, model_name, "url")
 
     # Determine if configuring an embedding model
     try:
@@ -72,10 +101,10 @@ async def get_client(
     except (AttributeError, KeyError):
         embedding = False
 
-    # Model Classes
+    # schema.Model Classes
     model_classes = {}
     if not embedding:
-        logger.debug("Configuring LL Model")
+        logger.debug("Configuring LL schema.Model")
         ll_common_params = {}
         for key in [
             "temperature",
@@ -87,11 +116,11 @@ async def get_client(
         ]:
             try:
                 logger.debug("--> Setting: %s; was sent %s", key, model_config[key])
-                ll_common_params[key] = model_config[key] or await get_key_value(model_objects, model_name, key)
+                ll_common_params[key] = model_config[key] or get_key_value(model_objects, model_name, key)
             except KeyError:
                 # Mainly for embeddings
                 continue
-        logger.debug("LL Model Parameters: %s", ll_common_params)
+        logger.debug("LL schema.Model Parameters: %s", ll_common_params)
         model_classes = {
             "OpenAI": lambda: ChatOpenAI(model=model_name, api_key=model_api_key, **ll_common_params),
             "CompatOpenAI": lambda: ChatOpenAI(
@@ -109,7 +138,7 @@ async def get_client(
             ),
             "ChatOCIGenAI": lambda oci_cfg=oci_config: ChatOCIGenAI(
                 model_id=model_name,
-                client=init_genai_client(oci_cfg),
+                client=oci.init_genai_client(oci_cfg),
                 compartment_id=oci_cfg.compartment_id,
                 model_kwargs={
                     (k if k != "max_completion_tokens" else "max_tokens"): v
@@ -119,7 +148,7 @@ async def get_client(
             ),
         }
     if embedding:
-        logger.debug("Configuring Embed Model")
+        logger.debug("Configuring Embed schema.Model")
         model_classes = {
             "OpenAIEmbeddings": lambda: OpenAIEmbeddings(model=model_name, api_key=model_api_key),
             "CompatOpenAIEmbeddings": lambda: OpenAIEmbeddings(
@@ -133,7 +162,7 @@ async def get_client(
             "HuggingFaceEndpointEmbeddings": lambda: HuggingFaceEndpointEmbeddings(model=model_url),
             "OCIGenAIEmbeddings": lambda oci_cfg=oci_config: OCIGenAIEmbeddings(
                 model_id=model_name,
-                client=init_genai_client(oci_cfg),
+                client=oci.init_genai_client(oci_cfg),
                 compartment_id=oci_cfg.compartment_id,
             ),
         }
@@ -147,7 +176,7 @@ async def get_client(
         else:
             logger.debug("Searching for %s in %s", model_api, model_classes)
             client = model_classes[model_api]()
-            logger.debug("Model Client: %s", client)
+            logger.debug("schema.Model Client: %s", client)
         return client
     except (UnboundLocalError, KeyError):
         logger.error("Unable to find client; expect trouble!")
