+++ src/server/api/v1/models.py
@@ -1,0 +1,89 @@
+"""
+Copyright (c) 2024, 2025, Oracle and/or its affiliates.
+Licensed under the Universal Permissive License v1.0 as shown at http://oss.oracle.com/licenses/upl.
+"""
+# spell-checker:ignore selectai
+
+from typing import Optional, get_args
+from fastapi import APIRouter, HTTPException, Query
+from fastapi.responses import JSONResponse
+
+from server.api.core import models
+import common.schema as schema
+import common.logging_config as logging_config
+
+logger = logging_config.logging.getLogger("endpoints.v1.models")
+
+auth = APIRouter()
+
+
+@auth.get("/v1/models/api", description="Get support model APIs", response_model=list)
+async def models_list_api(
+    model_type: Optional[schema.ModelTypeType] = Query(None),
+) -> list[schema.Model]:
+    """List all models APIs after applying filters if specified"""
+    logger.debug("Received models_list_api - type: %s", model_type)
+    if model_type == "ll":
+        return list(get_args(schema.LlAPI))
+    elif model_type == "embed":
+        return list(get_args(schema.EmbedAPI))
+    else:
+        return list()
+
+
+@auth.get("/v1/models", description="Get all models", response_model=list[schema.Model])
+async def models_list(
+    model_type: Optional[schema.ModelTypeType] = Query(None),
+) -> list[schema.Model]:
+    """List all models after applying filters if specified"""
+    logger.debug("Received models_list - type: %s", model_type)
+    models_ret = models.get_model(model_type=model_type)
+
+    return models_ret
+
+
+@auth.get("/v1/models/{name:path}", description="Get a single model", response_model=schema.Model)
+async def models_get(name: schema.ModelNameType) -> schema.Model:
+    """List a specific model"""
+    logger.debug("Received models_get - name: %s", name)
+
+    try:
+        models_ret = models.get_model(model_name=name)
+    except ValueError as ex:
+        raise HTTPException(status_code=404, detail=str(ex)) from ex
+
+    return models_ret[0]
+
+
+@auth.patch("/v1/models/{name:path}", description="Update a model", response_model=schema.Model)
+async def models_update(name: schema.ModelNameType, payload: schema.Model) -> schema.Model:
+    """Update a model"""
+    logger.debug("Received models_update - name: %s; payload: %s", name, payload)
+
+    model_upd = models.get_model(model_name=name)
+    for key, value in payload:
+        if hasattr(model_upd, key):
+            setattr(model_upd, key, value)
+        else:
+            raise HTTPException(status_code=404, detail=f"Model: Invalid setting - {key}.")
+
+    return models.get_model(model_name=name)
+
+
+@auth.post("/v1/models", description="Create a model", response_model=schema.Model)
+async def models_create(payload: schema.Model) -> schema.Model:
+    """Update a model"""
+    logger.debug("Received model_create - payload: %s", payload)
+
+    try:
+        return models.create_model(payload)
+    except ValueError as ex:
+        raise HTTPException(status_code=409, detail=str(ex)) from ex
+
+
+@auth.delete("/v1/models/{name:path}", description="Delete a model")
+async def models_delete(name: schema.ModelNameType) -> JSONResponse:
+    """Delete a model"""
+    logger.debug("Received models_delete - name: %s", name)
+    models.delete_model(name)
+    return JSONResponse(status_code=200, content={"message": f"Model: {name} deleted."})
